Move FakeCanary to common_util
Couldn't bring myself to move it back to test_storage.
